requirements.txt
bidict==0.22.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
plotly==5.17.0
pyaml==23.9.7
python-engineio==4.8.0
python-socketio==5.10.0
PyYAML==6.0.1
requests==2.31.0
simple-websocket==1.0.0
tenacity==8.2.3
urllib3==2.0.7
Werkzeug==3.0.0
wsproto==1.2.0
zipp==3.17.0
